ara-bidstudio-collection/OpenContracts/requirements/base.txt
pytz>=2024.2 # https://github.com/stub42/pytz
python-slugify==8.0.4  # https://github.com/un33k/python-slugify
Pillow>=10.4.0  # https://github.com/python-pillow/Pillow
argon2-cffi==25.1.0  # https://github.com/hynek/argon2_cffi
hiredis==3.2.1  # https://github.com/redis/hiredis-py
redis==5.2.1  # https://github.com/redis/redis-py
celery==5.5.3  # pyup: < 6.0  # https://github.com/celery/celery
flower==2.0.1  # https://github.com/mher/flower
django-celery-beat==2.6.0  # https://github.com/celery/django-celery-beat
pyjwt==2.10.1  # https://github.com/jpadilla/pyjwt
cryptography==45.0.6  # https://github.com/pyca/cryptography
pydantic==2.*
typing-extensions==4.*  # https://github.com/python/typing_extensions
requests==2.*  # https://requests.readthedocs.io/en/latest/
tokenizers>=0.21,<0.23  # Pin to prevent conflicts with transformers


# Django
# ------------------------------------------------------------------------------
django==4.2.24  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.12.0  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-crispy-forms==2.4  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2025.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-redis==6.0.0  # https://github.com/jazzband/django-redis
django-filter==24.3  # https://github.com/carltongibson/django-filter
django-storages[boto3,google]==1.14.6  # https://github.com/jschneier/django-storages
django-extensions==4.1  # https://github.com/django-extensions/django-extensions
whitenoise==6.7.0
django-tree-queries[admin]==0.21.0  # https://pypi.org/project/django-tree-queries/0.21.0/
django-cte==2.0.0  # https://github.com/dimagi/django-cte
pgvector  # For vector embeddings
channels==4.3.1  # https://github.com/django/channels
daphne==4.1.2  # https://github.com/django/daphne
channels-redis==4.3.0
Twisted[tls,http2]

# Django REST Framework
# ------------------------------------------------------------------------------
djangorestframework==3.16.1  # https://github.com/encode/django-rest-framework
django-cors-headers==4.9.0  # https://github.com/adamchainz/django-cors-headers
drf-extra-fields==3.7.0  # https://github.com/Hipo/drf-extra-fields

# Doc Analysis (NLP dependencies, placeholder) - WARNING, these are all in same environment... not ideal.
# ------------------------------------------------------------------------------
PyPDF2==3.0.1  # https://github.com/mstamy2/PyPDF2
pandas
plasmapdf==0.1.3  # https://github.com/Jsv4/plasmapdf
pdf2image
tiktoken==0.12.0  # https://github.com/openai/tiktoken
openai==1.102.0  # https://github.com/openai/openai-python
pytesseract
pydantic-ai==0.2.*
spacy

# Data Processing Tools
# -------------------------------------------------------------------------------
opencv-python==4.12.0.88 # https://github.com/opencv/opencv-python
filetype==1.2.0  # https://github.com/h2non/filetype.py

# Permissioning
# ------------------------------------------------------------------------------
django-guardian

# GraphQL
# ------------------------------------------------------------------------------
graphene-django==3.2.2  # TODO - upgrade at some point. Will require some refactoring.
django-graphql-jwt==0.4.0
django-ratelimit==4.1.0  # https://github.com/jsocol/django-ratelimit

# Telemetry
# ------------------------------------------------------------------------------
posthog==3.8.3  # https://github.com/posthog/posthog-python

# JSON Schema
# ------------------------------------------------------------------------------
jsonschema==4.25.1

# Not directly required, pinned by Snyk to avoid a vulnerability
# ------------------------------------------------------------------------------
twisted>=24.7.0rc1 # not directly required, pinned by Snyk to avoid a vulnerability
ipython>=8.10.0 # not directly required, pinned by Snyk to avoid a vulnerability
requests>=2.32.2 # not directly required, pinned by Snyk to avoid a vulnerability
setuptools>=70.0.0 # not directly required, pinned by Snyk to avoid a vulnerability
sqlparse>=0.5.0 # not directly required, pinned by Snyk to avoid a vulnerability
tornado>=6.4.2 # not directly required, pinned by Snyk to avoid a vulnerability
urllib3>=2.2.2 # not directly required, pinned by Snyk to avoid a vulnerability
zipp>=3.19.1 # not directly required, pinned by Snyk to avoid a vulnerability
pillow>=10.3.0 # not directly required, pinned by Snyk to avoid a vulnerability
